feat(): update plan item
